[polyselect/polyselect2l] codes for 32-bit machine and fix some warnings

+++ polyselect/polyselect2l_arith.h
@@ -7,8 +7,8 @@
 
 unsigned long invert (unsigned long, unsigned long);
 
-void roots_lift (unsigned long *, mpz_t, unsigned long,
-                 mpz_t, unsigned long, unsigned int);
+void roots_lift (uint64_t*, mpz_t, unsigned long,
+                 mpz_t, unsigned long, unsigned long int);
 
 void first_comb (unsigned long, unsigned long *);
 
@@ -23,8 +23,7 @@ void comp_sq_roots (header_t, qroots_t);
 
 void crt_sq (mpz_t, mpz_t, unsigned long *, unsigned long *);
 
-unsigned long return_q_rq (qroots_t, unsigned long *idx_q,
-                                  unsigned long k, mpz_t qqz,
-                                  mpz_t rqqz);
+uint64_t return_q_rq (qroots_t, unsigned long *idx_q,
+		      unsigned long k, mpz_t qqz, mpz_t rqqz);
 
 #endif
